"Completed" Some Stuff
“Completed” internships and mentorships as best as I could do without
talking to the client. More is probably to be done but for now it’s
looking fine. Also the capabilities to filter by ‘focus’ field is done.
It would be cool to have multiple focuses but right now each project
gets one!
Began working on the login screen and securing it properly.
Began events but I might need a little bit of direction to complete
this one. Though I could probably do a really good rough in the time
being.