Progress into Admin.html
Some progress into the admin dashboard. The idea is that they should be
able to see everything going on in the system through this dashboard.
Users, internships/mentorships, events (eventually). This will leave
things like UserStatus and whatever else I make specific type fields be
abstracted away. In the end, the website admins should be concerned
with the statuses not the underlying specific fields.