Cleaning Up, Preping New Stuff
Cleaned up a lot of the Servlet code to make it a lot more general and
easier to work with. All ajax calls should be improved on as should the
out messages should probably be normalized.
More stuff to fix but there’s some good stuff happening. Also
mentorships can be accepted and internships have the code to be
accepted but don’t pop up in their tab yet….