Lots of Stuff
There’s a lot of changes but it’s generally just quality of life
changes for both user and the codebase itself.

Other than that, the biggest change here is the internship and
mentorship information pages. This will allow (right now just the
admin) users to view the internship/mentorship itself instead of
viewing brief details that are presented in the main portions of the
site.