3 Things
Password Hasing - before we stored plain text passwords. Super awful.
But now we hash the passwords. Albeit simply but at least the passwords
aren’t just sitting in the database. Created a hashRunner that displays
a hash of an input. Useful for if manually inputting a user into the
server.

Accepting and Declining - We can now accept and decline users,
mentorships, and internships. Simple updating of database fields.

Session Saving - This was hard to get working but we got it now. The
server saves the user in req.session.user. This can be used to
determine user type, name, among other fields. All saved in the session
on the server side.